chore: update TopicPage and GalleryPage
