+++ client/src/components/Topic/TopicPage.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useContext } from 'react';
+import { useState, useEffect, useContext, useMemo } from 'react';
 import { Helmet } from 'react-helmet';
 import axios from 'axios';
 import TopicDetails from './TopicDetails';
@@ -24,8 +24,9 @@ export default function TopicPage() {
       .catch((err) => console.log(err));
   }, [refreshTopics]);
 
-  const TopicsSortedByDate = topics.sort(
-    (a, b) => new Date(b.createdAt) - new Date(a.createdAt),
+  const TopicsSortedByDate = useMemo(
+    () => topics.sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt)),
+    [topics],
   );
 
   const handleClickTopic = (topicId) => {
